Create floating menu component
As an intermediate step towards implementing menus, this adds a component for the variant of menus used on desktop. To make it ready for external use, I still need to implement the drawer menu variant used on mobile, and then package it up in a more high-level component that also handles menu triggers.