Refactor IndicatorIcon out of IconButton
So we have an icon-with-indicator component that we cvan use without
it being a button.

This also updates the possible colours for the indicator which we need for
the new threads notification work. I've renamed them to match the colours
they use rather than 'highlight' as previously because I found it the least
confusing. This does mean that this is a BREAKING CHANGE (although we could
add 'highlight' for backwards compat, I don't think it's necessary?)