Change how we align the chevron on a MenuItem
Hopefully the comment explains. The chevron SVG has a lot of blank
space and so needs shoving over to get it close enough to the right
hand edge. We did this with a negative margin but this causes some
of the chevron element to lie outside the parent which can make
scrollbars appear and/or cause it to cover the focus-active border.

This adjusts the viewBox of the SVG so it can just be the right size
for the space.