Bring back the tooltip warm-up period
To make it easier to browse through tooltips on neighboring elements, a common UX optimization is to allow the hover delay to be "warmed up", so that it goes to zero after the first hover. We used to have this before the tooltip component was rewritten.

This is a breaking change because consumers are now required to wrap their application in a <TooltipProvider> component once again.