Make tooltips more touchscreen-friendly
The current behavior on touchscreens is that tooltips will open as soon as you touch the trigger and remain open indefinitely due to focus and simulated hover. This feels unexpected, particularly on mobile.

The solution implemented here is to ignore non-visible focus and simulated hover events, and instead allow tooltips to be opened by long-pressing on the trigger. This matches the behavior of native tooltips on Android (though I don't know whether iOS has this same interaction pattern).