Improve accessibility of toggle and checkbox menu items
MenuItems can be used either inside a Menu component (which creates a roving tab index) or simply with an element of role menu. In the latter case, focus would previously land on the cosmetic <input> element, which is a bad idea as it's hidden from accessibility technologies. By turning the parent <div role="menuitemcheckbox"> into a <button> instead, we can ensure that focus always lands on the parent element. This is possible because ToggleMenuItem and CheckboxMenuItem are controlled components these days.